Small bug fix and enhanced logging.
There was a bug in the calculation of the target of the Q-functions 
(used in the Q-losses). The line used to be the following:

self.reward_scale * rewards * (1.0 - terminals) * self.discount_factor * 
value_next_observation

It has been changed to this:

self.reward_scale * rewards + (1.0 - terminals) * self.discount_factor * 
value_next_observation

Notice that one of the mlitiplier operators has been changed to 
addition.

Also logs alpha loss to Tensorboard and makes sure to log the average 
losses instead of the loss from the last gradient step.